+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 - Fix typo in params that contained `intencity` renaming it to `intensity`.
+- Fix bug `playOnce` and `playOnceOther` when call again without finish the last.
 
 ## 3.0.10
 - Create mixin `MovePerCell`.