+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix typo in params that contained `intencity` renaming it to `intensity`.
 - Fix bug `playOnce` and `playOnceOther` when call again without finish the last.
 - `BarLifeComponent` improvements. Now it center horizontally automatic
+- `Jumper` improvements.
 
 ## 3.0.10
 - Create mixin `MovePerCell`.